Delete Case_statement_to display_message_based_on space_available_in_the_system.pdf
Question 1: Explain the following bash script:
#!/bin/bash space_free=$( df -h | awk '{ print $5 }' | sort -n | tail -n 1 | sed 's/%//' ) case $space_free in
[1-5]*) echo Plenty of disk space available
;;
[6-7]*) echo There could be a problem in the near future
;;
8*) echo Maybe we should look at clearing out old files
;;
9*) echo We could have a serious problem on our hands soon
;;
*) echo Something is not quite right here
;;
esac
Solution: This script is a simple bash script that checks the amount of free disk space on the system and prints a message based on the percentage of free space.
'df' command is used to display information about the file system's disk space usage
'|' symbol pipes the output to the next command
'awk '{ print $5 }'' to extract the 5th column, which is the percentage of used space.
'sort –n' The output is then passed to sort the output numerically
'tail -n1' is used to get the last (highest) value.
'sed' command is used to remove the percentage sign from the output.
The resulting value is stored in the space_free variable.

Next, the script uses a case statement to check the value of space_free and print a message based on the range of values.
For example, if the value is between 1 and 5 (inclusive), it will print "Plenty of disk space available". If the value is between 6 and 7 (inclusive), it will print "There could be a problem in the near future". If the value is 8, it will print "Maybe we should look at clearing out old files", and so on. The * in the case statements is a wildcard that matches any value. The ;; at the end of each case indicates the end of the case statement. Finally, the esac keyword indicates the end of the case statement.